Give each class a separate header; hide QJsonItem behind include/details folder
